requirements.txt
pandas==0.25.3
mlflow==1.18.0
torch==1.7.1
plotly==4.5.4
wordcloud==1.8.1
htbuilder==0.5.0
streamlit==0.83.0
emoji==1.2.0
numpy==1.17.4
matplotlib==3.3.1
Pillow==8.2.0
